# Performance optimizations for Earley parser