… t/optimization/test-left-recursion-performance.t
… t/optimization/test-left-recursion.t
… t/optimization/test-leo-items.t
… t/optimization/test-memoization.t
… t/optimization/test-nullability-optimization.t
… t/optimization/test-nullability.t
… t/optimization/test-prediction-memoization.t