## Summary
This PR implements several targeted performance optimizations for the chalk Earley parser, resulting in a 28% runtime improvement (from 121s to 87s) while maintaining 100% parsing success.

• **Regex caching**: Added state-based caching in `terminal_to_regex()` to avoid recompiling identical regexes
• **Computed value caching**: Cache nullable rule computations and item properties to reduce redundant calculations  
• **Rule consolidation**: Inline single-character terminal rules and consolidate terminal-only operator rules using regex patterns
• **Data structure optimizations**: Use array indexing instead of hash lookups for position-based chart access, cache rule properties in item fields
• **Algorithm cleanup**: Remove unused semiring implementations (Boolean, Viterbi) and obsolete optimization tests

## Test plan
- [x] Verify parsing completes at 100% success rate on chalk-grammar.pl
- [x] Confirm runtime improvement from 121s baseline to 87s (28% faster)
- [x] Ensure all existing functionality preserved
- [x] Test STDIN input handling with main execution block